feat/LIVE-2613 Moved helper to excluded jest folder
