+++ libs/ledger-live-common/src/__tests__/api/test-helpers/scenarios.ts
