+++ libs/ledgerjs/packages/errors/src/index.ts
@@ -298,8 +298,6 @@ export const StatusCodes = {
   CONTRADICTION_INVALIDATION: 0x9810,
   CONTRADICTION_SECRET_CODE_STATUS: 0x9808,
   CUSTOM_IMAGE_BOOTLOADER: 0x662f,
-  CUSTOM_IMAGE_BOOTLOADER: 0x662f,
-  CUSTOM_IMAGE_EMPTY: 0x662e,
   CUSTOM_IMAGE_EMPTY: 0x662e,
   FILE_ALREADY_EXISTS: 0x6a89,
   FILE_NOT_FOUND: 0x9404,
